refactor: move pinata service to services folder
